+++ js/app/engineClient.js
@@ -70,6 +70,8 @@
   EngineClient.prototype.diffSummary = function (stepId) { return this.send({ type: 'diffSummary', stepId: stepId }); };
   EngineClient.prototype.search = function (stepId, query, limit) { return this.send({ type: 'search', stepId: stepId, query: query, limit: limit }); };
   EngineClient.prototype.findRows = function (stepId, lookup, limit) { return this.send({ type: 'findRows', stepId: stepId, lookup: lookup, limit: limit }); };
+  EngineClient.prototype.batch = function (file, options, steps, output) { return this.send({ type: 'batch', file: file, options: options, steps: steps, output: output }); };
+  EngineClient.prototype.zip = function (entries) { return this.send({ type: 'zip', entries: entries }); };
   EngineClient.prototype.exportStep = function (stepId, options) { return this.send({ type: 'export', stepId: stepId, options: options }); };
 
   DL.EngineClient = EngineClient;